+++ flexmeasures/conftest.py
@@ -3,6 +3,7 @@ import pytest
 from random import random
 from datetime import datetime, timedelta
 from typing import List, Dict
+import pytz
 
 from isodate import parse_duration
 import pandas as pd
@@ -494,14 +495,18 @@ def add_market_prices(db: SQLAlchemy, setup_assets, setup_markets, setup_sources
 
     # one day of test data (one complete sine curve)
     time_slots = pd.date_range(
-        datetime(2015, 1, 1), datetime(2015, 1, 2), freq="1H", closed="left"
+        datetime(2015, 1, 1),
+        datetime(2015, 1, 2),
+        freq="1H",
+        closed="left",
+        tz="Europe/Amsterdam",
     )
     values = [
         random() * (1 + np.sin(x * 2 * np.pi / 24)) for x in range(len(time_slots))
     ]
     day1_beliefs = [
         TimedBelief(
-            event_start=as_server_time(dt),
+            event_start=dt,
             belief_horizon=timedelta(hours=0),
             event_value=val,
             source=setup_sources["Seita"],
@@ -513,12 +518,16 @@ def add_market_prices(db: SQLAlchemy, setup_assets, setup_markets, setup_sources
 
     # another day of test data (8 expensive hours, 8 cheap hours, and again 8 expensive hours)
     time_slots = pd.date_range(
-        datetime(2015, 1, 2), datetime(2015, 1, 3), freq="1H", closed="left"
+        datetime(2015, 1, 2),
+        datetime(2015, 1, 3),
+        freq="1H",
+        closed="left",
+        tz="Europe/Amsterdam",
     )
     values = [100] * 8 + [90] * 8 + [100] * 8
     day2_beliefs = [
         TimedBelief(
-            event_start=as_server_time(dt),
+            event_start=dt,
             belief_horizon=timedelta(hours=0),
             event_value=val,
             source=setup_sources["Seita"],
@@ -573,7 +582,7 @@ def create_test_battery_assets(
         max_soc_in_mwh=5,
         min_soc_in_mwh=0,
         soc_in_mwh=2.5,
-        soc_datetime=as_server_time(datetime(2015, 1, 1)),
+        soc_datetime=pytz.timezone("Europe/Amsterdam").localize(datetime(2015, 1, 1)),
         soc_udi_event_id=203,
         latitude=10,
         longitude=100,
@@ -591,7 +600,7 @@ def create_test_battery_assets(
         max_soc_in_mwh=5,
         min_soc_in_mwh=0,
         soc_in_mwh=2.5,
-        soc_datetime=as_server_time(datetime(2040, 1, 1)),
+        soc_datetime=pytz.timezone("Europe/Amsterdam").localize(datetime(2040, 1, 1)),
         soc_udi_event_id=203,
         latitude=10,
         longitude=100,
@@ -644,7 +653,7 @@ def add_charging_station_assets(
         max_soc_in_mwh=5,
         min_soc_in_mwh=0,
         soc_in_mwh=2.5,
-        soc_datetime=as_server_time(datetime(2015, 1, 1)),
+        soc_datetime=pytz.timezone("Europe/Amsterdam").localize(datetime(2015, 1, 1)),
         soc_udi_event_id=203,
         latitude=10,
         longitude=100,
@@ -662,7 +671,7 @@ def add_charging_station_assets(
         max_soc_in_mwh=5,
         min_soc_in_mwh=0,
         soc_in_mwh=2.5,
-        soc_datetime=as_server_time(datetime(2015, 1, 1)),
+        soc_datetime=pytz.timezone("Europe/Amsterdam").localize(datetime(2015, 1, 1)),
         soc_udi_event_id=203,
         latitude=10,
         longitude=100,