+++ flexmeasures/data/models/planning/charging_station.py
@@ -1,4 +1,4 @@
-from typing import Optional, Union
+from typing import List, Optional, Union
 from datetime import datetime, timedelta
 
 import pandas as pd
@@ -10,6 +10,7 @@ from flexmeasures.data.models.planning.utils import (
     initialize_series,
     add_tiny_price_slope,
     get_prices,
+    get_power_values,
     fallback_charging_policy,
 )
 
@@ -25,7 +26,10 @@ def schedule_charging_station(
     soc_max: Optional[float] = None,
     roundtrip_efficiency: Optional[float] = None,
     prefer_charging_sooner: bool = True,
-    price_sensor: Optional[Sensor] = None,
+    consumption_price_sensor: Optional[Sensor] = None,
+    production_price_sensor: Optional[Sensor] = None,
+    inflexible_device_sensors: Optional[List[Sensor]] = None,
+    belief_time: Optional[datetime] = None,
     round_to_decimals: Optional[int] = 6,
 ) -> Union[pd.Series, None]:
     """Schedule a charging station asset based directly on the latest beliefs regarding market prices within the specified time
@@ -53,13 +57,23 @@ def schedule_charging_station(
         soc_max = sensor.get_attribute("max_soc_in_mwh", max(soc_targets.values))
 
     # Check for known prices or price forecasts, trimming planning window accordingly
-    prices, (start, end) = get_prices(
+    up_deviation_prices, (start, end) = get_prices(
         (start, end),
         resolution,
-        price_sensor=price_sensor,
+        beliefs_before=belief_time,
+        price_sensor=consumption_price_sensor,
         sensor=sensor,
         allow_trimmed_query_window=True,
     )
+    down_deviation_prices, (start, end) = get_prices(
+        (start, end),
+        resolution,
+        beliefs_before=belief_time,
+        price_sensor=production_price_sensor,
+        sensor=sensor,
+        allow_trimmed_query_window=True,
+    )
+
     # soc targets are at the end of each time slot, while prices are indexed by the start of each time slot
     soc_targets = soc_targets.tz_convert("UTC")
     start = pd.Timestamp(start).tz_convert("UTC")
@@ -69,20 +83,23 @@ def schedule_charging_station(
     # Add tiny price slope to prefer charging now rather than later, and discharging later rather than now.
     # We penalise the future with at most 1 per thousand times the price spread.
     if prefer_charging_sooner:
-        prices = add_tiny_price_slope(prices, "event_value")
+        up_deviation_prices = add_tiny_price_slope(up_deviation_prices, "event_value")
+        down_deviation_prices = add_tiny_price_slope(
+            down_deviation_prices, "event_value"
+        )
 
     # Set up commitments to optimise for
     commitment_quantities = [initialize_series(0, start, end, resolution)]
 
     # Todo: convert to EUR/(deviation of commitment, which is in MW)
     commitment_upwards_deviation_price = [
-        prices.loc[start : end - resolution]["event_value"]
+        up_deviation_prices.loc[start : end - resolution]["event_value"]
     ]
     commitment_downwards_deviation_price = [
-        prices.loc[start : end - resolution]["event_value"]
+        down_deviation_prices.loc[start : end - resolution]["event_value"]
     ]
 
-    # Set up device constraints (only one device for this EMS)
+    # Set up device constraints: only one scheduled flexible device for this EMS (at index 0), plus the forecasted inflexible devices (at indices 1 to n).
     columns = [
         "equals",
         "max",
@@ -91,7 +108,18 @@ def schedule_charging_station(
         "derivative max",
         "derivative min",
     ]
-    device_constraints = [initialize_df(columns, start, end, resolution)]
+    if inflexible_device_sensors is None:
+        inflexible_device_sensors = []
+    device_constraints = [initialize_df(columns, start, end, resolution)] * (
+        1 + len(inflexible_device_sensors)
+    )
+    for i, inflexible_sensor in enumerate(inflexible_device_sensors):
+        device_constraints[i + 1]["derivative equals"] = get_power_values(
+            query_window=(start, end),
+            resolution=resolution,
+            beliefs_before=belief_time,
+            sensor=inflexible_sensor,
+        )
     device_constraints[0]["equals"] = soc_targets.shift(-1, freq=resolution).values * (
         timedelta(hours=1) / resolution
     ) - soc_at_start * (
@@ -121,9 +149,13 @@ def schedule_charging_station(
     device_constraints[0]["derivative down efficiency"] = roundtrip_efficiency**0.5
     device_constraints[0]["derivative up efficiency"] = roundtrip_efficiency**0.5
 
-    # Set up EMS constraints (no additional constraints)
+    # Set up EMS constraints
     columns = ["derivative max", "derivative min"]
     ems_constraints = initialize_df(columns, start, end, resolution)
+    ems_capacity = sensor.generic_asset.get_attribute("capacity_in_mw")
+    if ems_capacity is not None:
+        ems_constraints["derivative min"] = ems_capacity * -1
+        ems_constraints["derivative max"] = ems_capacity
 
     ems_schedule, expected_costs, scheduler_results = device_scheduler(
         device_constraints,